// src/app/common/enums.ts
export enum Names {
  ExhibitionFormTitle = '闇市出品フォーム',
  Exhibit = '出品',
}

export enum IDs {
  Create = 'create',
}

export enum ExhibitElementNames {
  Name = '商品名',
  Description = '商品説明',
  StartValue = '開始価格',
  EndValue = '即決価格',
  ExhibitionPeriod = '出品期間',
}

export enum ExhibitElementIDs {
  Name = 'name',
  Description = 'description',
  StartValue = 'start_value',
  EndValue = 'end_value',
  ExhibitionPeriod = 'exhibition_period',
}